requirements.txt
Flask==3.1.1
gunicorn==23.0.0
PyYAML==6.0.2
Flask-Login==0.6.3
Pillow>=11.1.0
Flask-WTF==1.2.2
Flask-Babel==4.0.0
nh3==0.2.21
